exportGraphdata method created
